appimage move to realse and refactor

+++ tools/app-image-creator/buildset-teardown.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 cd "$(dirname "$0")"
 
-rm -rf ./buildset
+if [[ -d ./AppDir ]]; then
+    rm -rf ./AppDir
+fi
+
+if [[ -d ./buildset ]]; then
+    rm -rf ./buildset
+fi
+
 rm -rf ./build
